Allow embedding with markdown and pandoc
certain syntax elements are defined as `contained=TOP`. However, when
embedding the zsh syntax into another language like pandoc or markdown
does, this will prevent to match those syntax items, because the top level
items will then be used for the top level syntax language
markdown/pandoc and not zsh.

Therefore, check if the @zsh or @markdownHighlightzsh cluster is defined
and if so, instead of defining `contains=TOP` use the correct
`contains=@<cluster>` (which should be the cluster defined for the
embedded lanuage, e.g. either @zsh for pandoc or @markdownHighlightzsh
cluster for markdown).

closes #21