Remove leading , from b:match_words
For reasons I have not yet elucidated, this leading "," causes "%" to
behave wrongly when placed before a bracket or match. That is, given the
document below (with &filetype=zsh):

  f () {
    :
  }

If you place the cursor on "f" and press "%", you should expect to jump
to the matching parentheses. With the leading "," this behavior is
broken and no match is made. Removing it corrects the behavior.