ftplugin: Undefine MANPAGER before running run-help
fixes #25

Apparently, if MANPAGER is defined and run-help is used (and piped
through less) this makes vim totally unresponsive.

So fix it, by making sure MANPAGER is undefined before calling out to
the run-help command. It should be noted, that is should only be
necessary when run-help is piped through less, therefore, it should not
be needed for the :terminal version.