Improve performance of zshOption
The syntax file was quite slow in my zshrc; in some places in the file
it was so slow that it's like typing over a slow ssh connection. Turns
out much of it is in the zshOption regexp, which has a lot of needless
grouping. This rewrites it to use a single %(...)

Comparison of performance with my benchmark script:

Before

	TOTAL      COUNT  MATCH   SLOWEST     AVERAGE   NAME         PATTERN
	33.071598   2816   1410   0.025515    0.011744  zshOption    \%(\%(\<no_..

After

	TOTAL      COUNT  MATCH   SLOWEST     AVERAGE   NAME         PATTERN
	2.476799   2715    1309   0.001910    0.000912  zshOption    \v <%(no_?)..

Still not winning any performance prizes here, but it's a lot better.

This also adds the optional underscores feature: the options
"hist_ignore_dups" "histignore_dups" "hist_ignoredups" "histignoredups"
are all equal as far as zsh is concerned, with or without "no" or "no_"
prefix. This seems to have negligible performance impact.