Add a few tests
As a bit of a follow-up to #42, this adds tests for the syntax
highlighting.

Example zsh files are in testdata/*.zsh; after writing test files are
verifying it's all good you can generate with "make update-tests", and
you can run it with "make test". I included the testing.vim as a
submodule here.