Link zshOperator to Operator
This highlights ZSH operators like && as operators.

A long time ago it looked like there was a thought that either operators
or redirection should be highlighted as operators, with redirection
getting linked in the end (9215b31). However, these are not mutually
exclusive. It's also uncommon to have an operator adjacent to a
redirection, so there's no cause for confusion.